+++ app/Game/Werewolf/Command/Status.hs
@@ -25,11 +25,12 @@ import Data.List
 import Data.Text (Text)
 
 -- TODO (hjw): remove Message.Engine
-import Game.Werewolf                 hiding (getPendingVoters)
-import Game.Werewolf.Command
-import Game.Werewolf.Message.Command
-import Game.Werewolf.Message.Engine
-import Game.Werewolf.Util
+import           Game.Werewolf                 hiding (getPendingVoters)
+import           Game.Werewolf.Command
+import           Game.Werewolf.Message.Command
+import           Game.Werewolf.Message.Engine
+import qualified Game.Werewolf.Role            as Role
+import           Game.Werewolf.Util
 
 circleCommand :: Text -> Bool -> Command
 circleCommand callerName includeDead = Command $ do
@@ -64,7 +65,7 @@ pingRole role' = do
     tell [pingPlayerMessage $ player ^. name]
     where
         pingRoleMessage game
-            | has (activity . _Diurnal) role'   = pingDiurnalRoleMessage role'
+            | has (Role.activity . _Diurnal) role'   = pingDiurnalRoleMessage role'
             | otherwise                         = pingNocturnalRoleMessage role' game
 
 pingVillagers :: (MonadState Game m, MonadWriter [Message] m) => m ()
@@ -86,7 +87,12 @@ statusCommand :: Text -> Command
 statusCommand callerName = Command $ do
     game <- get
 
-    tell [ currentStageMessage callerName game
+    tell [ currentStageMessage game
         , rolesInGameMessage (Just callerName) game
         , playersInGameMessage callerName game
         ]
+    where
+        currentStageMessage game
+            | has (stage . _GameOver) game              = gameIsOverMessage callerName
+            | has (stage . activity . _Diurnal) game    = currentDiurnalTurnMessage callerName game
+            | otherwise                                 = currentNocturnalTurnMessage callerName game